+++ package.json
@@ -18,6 +18,7 @@
     "@reduxjs/toolkit": "^1.9.6",
     "@rneui/themed": "^4.0.0-rc.8",
     "@shopify/restyle": "^2.4.2",
+    "@types/humanize-plus": "^1.8.0",
     "axios": "^1.5.0",
     "eas-cli": "^5.2.0",
     "expo": "~48.0.10",
@@ -30,6 +31,7 @@
     "expo-updates": "~0.16.4",
     "formik": "^2.4.5",
     "global": "^4.4.0",
+    "humanize-plus": "^1.8.2",
     "moment": "^2.29.4",
     "nativewind": "^2.0.11",
     "postcss": "8.4.23",
