+++ src/Home/QuoteConfirm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./QuoteConfirm"
